Ai update
The ai now turns to a random angle when a specific distance away from
the camera for the wander state, it always stays within a specific range
of the camera also to ensure that the ai doesn't wander away from the
player. when the camera is fairly close to the ai object it then moves
towards the camera and stops when colliding with it. next I need to add
movement relative to the rotation of the object next so when in the
wander state it moves forward depending on the direction its facing.